Move CommunityPipeBuilder to the interpreted-runtime module
