+++ community/cypher/cypher/src/main/scala/org/neo4j/cypher/internal/compatibility/v3_5/runtime/PipeExecutionPlanBuilder.scala
@@ -22,8 +22,9 @@ package org.neo4j.cypher.internal.compatibility.v3_5.runtime
 import org.neo4j.cypher.internal.compatibility.v3_5.runtime.executionplan._
 import org.neo4j.cypher.internal.ir.v3_5.PeriodicCommit
 import org.neo4j.cypher.internal.planner.v3_5.spi.TokenContext
+import org.neo4j.cypher.internal.runtime.interpreted.CommunityPipeBuilder
 import org.neo4j.cypher.internal.runtime.interpreted.commands.convert.ExpressionConverters
-import org.neo4j.cypher.internal.runtime.interpreted.pipes.Pipe
+import org.neo4j.cypher.internal.runtime.interpreted.pipes.{Pipe, PipeBuilderFactory, PipeExecutionBuilderContext}
 import org.neo4j.cypher.internal.v3_5.logical.plans.{LogicalPlan, LogicalPlans, Limit => LimitPlan, LoadCSV => LoadCSVPlan, Skip => SkipPlan}
 
 class PipeExecutionPlanBuilder(pipeBuilderFactory: PipeBuilderFactory,
@@ -52,10 +53,4 @@ object CommunityPipeBuilderFactory extends PipeBuilderFactory {
            (implicit context: PipeExecutionBuilderContext, tokenContext: TokenContext): CommunityPipeBuilder = {
     CommunityPipeBuilder(recurse, readOnly, expressionConverters, recursePipes(recurse), tokenContext)(context.semanticTable)
   }
-}
-
-trait PipeBuilder extends LogicalPlans.Mapper[Pipe] {
-  override def onLeaf(plan: LogicalPlan): Pipe
-  override def onOneChildPlan(plan: LogicalPlan, source: Pipe): Pipe
-  override def onTwoChildPlan(plan: LogicalPlan, lhs: Pipe, rhs: Pipe): Pipe
 }
