+++ community/cypher/interpreted-runtime/src/main/scala/org/neo4j/cypher/internal/runtime/interpreted/pipes/PipeExecutionBuilderContext.scala
@@ -17,16 +17,11 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.neo4j.cypher.internal.compatibility.v3_5.runtime.executionplan.builders.prepare
+package org.neo4j.cypher.internal.runtime.interpreted.pipes
 
-import org.neo4j.cypher.internal.runtime.interpreted.commands.expressions.Expression
-import org.neo4j.cypher.internal.runtime.interpreted.commands.values.KeyToken
-import org.neo4j.cypher.internal.planner.v3_5.spi.TokenContext
+import org.neo4j.cypher.internal.frontend.v3_5.semantics.SemanticTable
+import org.neo4j.cypher.internal.planner.v3_5.spi.PlanningAttributes.{Cardinalities, ReadOnlies}
 
-object KeyTokenResolver {
-  /*this is what you should use!*/
-  def resolveExpressions(expr: Expression, ctx: TokenContext): Expression = expr match {
-    case (keyToken: KeyToken) => keyToken.resolve(ctx)
-    case _                    => expr
-  }
-}
+case class PipeExecutionBuilderContext(semanticTable: SemanticTable,
+                                       readOnlies: ReadOnlies,
+                                       cardinalities: Cardinalities)
